% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Concentration Data}
\format{
A data frame with 145 rows and 6 columns:
\describe{
  \item{SPECIES}{Fish species}
  \item{RIVER}{Name of the river}
  \item{MILE}{How far down river}
  \item{DDT}{Concentration of DDT [ppm]}
  \item{LENGTH}{Length of the fish [cm]}
  \item{WEIGHT}{Weight of the fish [g]}
}
}
\source{
Textbook example dataset from MATH 4753
}
\usage{
ddt
}
\description{
This dataset contains measurements related to DDT concentrations in fish species sampled from various rivers.
}
\keyword{datasets}
