% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIre damage data}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance the fire burnedd}
  \item{DAMAGE}{The amount of damage the fire caused}
}
}
\source{
Textbook example dataset from MATH 4753
}
\usage{
fire
}
\description{
Dataset with fire damage statistics
}
\keyword{datasets}
