% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{Sample size}

\item{p}{probability}
}
\value{
a barplot
}
\description{
mybin
}
\examples{
mybin(iter = 1000,n = 18, p = 0.3)
}
