% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrapping}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{function to that takes a numeric vector}

\item{alpha}{significance level of confidence interval}

\item{cx}{character expansion factor for text size in plots}

\item{...}{additional parameters}
}
\value{
A list
}
\description{
Bootstrapping
}
\examples{
#' \dontrun{
  data(ddt, package="MATH4753F25thayer")
  obj <- myboot2(x = ddt$DDT, fun="mean")
  obj$ci
}

}
