% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Uniform CLT}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{size of each iteration}

\item{iter}{iterations simulated}

\item{a}{lower bound for uniform distribution}

\item{b}{upper bound for uniform distribution}
}
\value{
A histogram
}
\description{
Uniform CLT
}
\examples{
mycltu(n=1,iter=10000)
}
