% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{List}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{area under the curve of which is shaded up to}
}
\value{
a list with mu, sigma, and probability of a
}
\description{
List
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)
}
