% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Optimal tickets sales}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of seats available}

\item{gamma}{Probability of overbooking}

\item{p}{Probability a passenger with a ticket shows}
}
\value{
Two plots and a list containing five components
}
\description{
Optimal tickets sales
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
